* Fixed spell emoting * Added relList for faster getClosestFreeTile working, this time works properly * Added extended position for getClosestFreeTile (+/-2) * Cleanups * Added doc/SIGNALS
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@617 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ map.cpp
@@ -215,7 +215,7 @@ bool Map::placeCreature(const Position& centerPos, Creature* creature, bool exte
 			foundTile = true;
 	}
 
-	typedef std::pair<int32_t, int32_t> relPair;
+	typedef std::pair<uint8_t, uint8_t> relPair;
 	uint8_t shufflePos = 0;
 
 	std::vector<relPair> relList;
